+++ src/rust_connection/mod.rs
@@ -5,8 +5,7 @@ use std::io::IoSlice;
 use std::sync::{Condvar, Mutex, MutexGuard, TryLockError};
 
 use crate::connection::{
-    compute_length_field, Connection, DiscardMode, ReplyOrError, RequestConnection, RequestKind,
-    SequenceNumber,
+    compute_length_field, Connection, ReplyOrError, RequestConnection, RequestKind,
 };
 use crate::cookie::{Cookie, CookieWithFds, VoidCookie};
 pub use crate::errors::{ConnectError, ConnectionError, ParseError, ReplyError, ReplyOrIdError};
@@ -15,23 +14,21 @@ use crate::protocol::bigreq::{ConnectionExt as _, EnableReply};
 use crate::protocol::xproto::{Setup, SetupRequest, GET_INPUT_FOCUS_REQUEST};
 use crate::utils::RawFdContainer;
 use crate::x11_utils::{ExtensionInformation, Serialize, TryParse, TryParseFd};
+use x11rb_protocol::connection::{Connection as ConnectionInner, PollReply, ReplyFdKind};
+use x11rb_protocol::{DiscardMode, RawEventAndSeqNumber, SequenceNumber};
 
 mod id_allocator;
-mod inner;
 mod packet_reader;
 mod parse_display;
 mod stream;
 mod write_buffer;
 mod xauth;
 
-use inner::PollReply;
 use packet_reader::PacketReader;
 pub use stream::{DefaultStream, PollMode, Stream};
 use write_buffer::WriteBuffer;
 
 type Buffer = <RustConnection as RequestConnection>::Buf;
-/// The raw bytes of an event received by [`RustConnection`] and its sequence number.
-pub type RawEventAndSeqNumber = crate::connection::RawEventAndSeqNumber<Buffer>;
 /// A combination of a buffer and a list of file descriptors for use by [`RustConnection`].
 pub type BufWithFds = crate::connection::BufWithFds<Buffer>;
 
@@ -42,14 +39,7 @@ enum MaxRequestBytes {
     Known(usize),
 }
 
-type MutexGuardInner<'a> = MutexGuard<'a, inner::ConnectionInner>;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub(crate) enum ReplyFdKind {
-    NoReply,
-    ReplyWithoutFDs,
-    ReplyWithFDs,
-}
+type MutexGuardInner<'a> = MutexGuard<'a, ConnectionInner>;
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub(crate) enum BlockingMode {
@@ -67,7 +57,7 @@ pub(crate) enum BlockingMode {
 /// to be buffered.
 #[derive(Debug)]
 pub struct RustConnection<S: Stream = DefaultStream> {
-    inner: Mutex<inner::ConnectionInner>,
+    inner: Mutex<ConnectionInner>,
     stream: S,
     // This mutex is only locked with `try_lock` (never blocks), so a simpler
     // lock based only on a atomic variable would be more efficient.
@@ -104,7 +94,7 @@ pub struct RustConnection<S: Stream = DefaultStream> {
 // lock on `inner` is released so that other threads can make progress. If more threads want to
 // read while `read` is already locked, they sleep on `reader_condition`. The actual reader will
 // then notify this condition variable once it is done reading.
-// 
+//
 // n.b. notgull: write_buffer follows the same rules
 //
 // The condition variable is necessary since one thread may read packets that another thread waits
@@ -182,14 +172,10 @@ impl<S: Stream> RustConnection<S> {
     /// It is assumed that `setup` was just received from the server. Thus, the first reply to a
     /// request that is sent will have sequence number one.
     pub fn for_connected_stream(stream: S, setup: Setup) -> Result<Self, ConnectError> {
-        Self::for_inner(stream, inner::ConnectionInner::new(), setup)
+        Self::for_inner(stream, ConnectionInner::new(), setup)
     }
 
-    fn for_inner(
-        stream: S,
-        inner: inner::ConnectionInner,
-        setup: Setup,
-    ) -> Result<Self, ConnectError> {
+    fn for_inner(stream: S, inner: ConnectionInner, setup: Setup) -> Result<Self, ConnectError> {
         let allocator =
             id_allocator::IdAllocator::new(setup.resource_id_base, setup.resource_id_mask)?;
         Ok(RustConnection {
@@ -278,13 +264,11 @@ impl<S: Stream> RustConnection<S> {
             let write_result = if !partial_buf.is_empty() {
                 // "inner" is held, passed into this function, so this should never be held
                 let mut write_buffer = self.write_buffer.try_lock().expect("inner is held");
-                write_buffer
-                    .write(&self.stream, partial_buf, &mut fds)
-            } else {                
+                write_buffer.write(&self.stream, partial_buf, &mut fds)
+            } else {
                 // same as above
                 let mut write_buffer = self.write_buffer.try_lock().expect("inner is held");
-                write_buffer
-                    .write_vectored(&self.stream, bufs, &mut fds)
+                write_buffer.write_vectored(&self.stream, bufs, &mut fds)
             };
             match write_result {
                 Ok(0) => {
@@ -332,9 +316,11 @@ impl<S: Stream> RustConnection<S> {
         &'a self,
         mut inner: MutexGuardInner<'a>,
     ) -> std::io::Result<MutexGuardInner<'a>> {
-        while inner.write_buffer.needs_flush() {
+        // n.b. notgull: inner guard is held
+        let mut write_buffer = self.write_buffer.try_lock().expect("inner is held");
+        while write_buffer.needs_flush() {
             self.stream.poll(PollMode::ReadAndWritable)?;
-            match inner.write_buffer.flush(&self.stream) {
+            match write_buffer.flush(&self.stream) {
                 // Flush completed
                 Ok(()) => break,
                 Err(ref e) if e.kind() == std::io::ErrorKind::WouldBlock => {
@@ -624,7 +610,9 @@ impl<S: Stream> RequestConnection for RustConnection<S> {
 }
 
 impl<S: Stream> Connection for RustConnection<S> {
-    fn wait_for_raw_event_with_sequence(&self) -> Result<RawEventAndSeqNumber, ConnectionError> {
+    fn wait_for_raw_event_with_sequence(
+        &self,
+    ) -> Result<RawEventAndSeqNumber<Vec<u8>>, ConnectionError> {
         let mut inner = self.inner.lock().unwrap();
         loop {
             if let Some(event) = inner.poll_for_event_with_sequence() {
@@ -636,7 +624,7 @@ impl<S: Stream> Connection for RustConnection<S> {
 
     fn poll_for_raw_event_with_sequence(
         &self,
-    ) -> Result<Option<RawEventAndSeqNumber>, ConnectionError> {
+    ) -> Result<Option<RawEventAndSeqNumber<Vec<u8>>>, ConnectionError> {
         let mut inner = self.inner.lock().unwrap();
         if let Some(event) = inner.poll_for_event_with_sequence() {
             Ok(Some(event))
